requirements.txt
tensorflow>=2.10
flask
numpy
Pillow